Move toggle notify
