+++ src/app/containers/Labels/Container.js
@@ -1,30 +1,14 @@
 import React, { useState, useEffect, useContext } from 'react';
 import { c } from 'ttag';
-import {
-    Title,
-    SubTitle,
-    LearnMore,
-    useApiResult,
-    Paragraph,
-    useApiWithoutResult,
-    useNotifications
-} from 'react-components';
+import { Title, SubTitle, LearnMore, useApiResult, Paragraph, useLabels, useApiWithoutResult } from 'react-components';
 import { arrayMove } from 'react-sortable-hoc';
-import { getLabels, orderLabels, updateLabel, createLabel } from 'proton-shared/lib/api/labels';
+import { orderLabels } from 'proton-shared/lib/api/labels';
 
 import LabelSortableList from '../../components/Labels/LabelSortableList';
 import ActionsLabelToolbar from '../../components/Labels/ActionsLabelToolbar';
 
 function LabelsContainer() {
-    const { result: { Labels = [] } = {}, loading } = useApiResult(getLabels, []);
-    const [list, setLabels] = useState(Labels);
-    const { createNotification } = useNotifications();
-
-    useEffect(() => {
-        setLabels(Labels);
-    }, [Labels]);
-
-    const updateRequest = useApiWithoutResult(updateLabel);
+    const [list, loading] = useLabels();
     const orderRequest = useApiWithoutResult(orderLabels);
 
     const sort = async (labels) => {
@@ -37,36 +21,6 @@ function LabelsContainer() {
         const list = arrayMove(Labels, oldIndex, newIndex);
         sort(list);
         setLabels(list);
-    };
-
-    const handleToggleChange = (label) => async () => {
-        const newLabel = {
-            ...label,
-            Notify: +!label.Notify
-        };
-        await updateRequest.request(label.ID, newLabel);
-        createNotification({
-            text: c('label/folder notification').t`${label.Name} updated`
-        });
-    };
-
-    const handleAdd = (label) => {
-        setLabels(list.concat(label));
-    };
-
-    const handleEditLabel = (label) => {
-        setLabels(
-            list.map((item) => {
-                if (label.ID === item.ID) {
-                    return label;
-                }
-                return item;
-            })
-        );
-    };
-
-    const handleRemoveLabel = ({ ID }) => {
-        setLabels(list.filter((label) => label.ID !== ID));
     };
 
     const getScrollContainer = () => document.querySelector('.main-area');
@@ -84,7 +38,7 @@ function LabelsContainer() {
                     <LearnMore url="https://protonmail.com" />
                 </p>
                 <nav className="mb1f flex">
-                    <ActionsLabelToolbar onAdd={handleAdd} />
+                    <ActionsLabelToolbar />
                 </nav>
 
                 {loading ? <div className="square-color bordered-container center" aria-busy="true" /> : null}
@@ -94,9 +48,6 @@ function LabelsContainer() {
                         getContainer={getScrollContainer}
                         pressDelay={200}
                         items={list}
-                        onEditLabel={handleEditLabel}
-                        onRemoveLabel={handleRemoveLabel}
-                        onToggleChange={handleToggleChange}
                         onSortEnd={onSortEnd}
                     />
                 ) : (
